ml-service/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.1
torchvision>=0.15.2

# Image Processing
Pillow>=10.0.1
opencv-python>=4.8.1.78
imageio>=2.31.1

# Data Processing
scipy>=1.11.1
matplotlib>=3.7.2
seaborn>=0.12.2

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.2

# Logging and monitoring
loguru>=0.7.2

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.9.1
flake8>=6.1.0